Blank display after 1 hour - it fades after 1000 hours
See note on Adafruit product page https://www.adafruit.com/product/938

Switch to representing time with a i64 to avoid the 50 day problem,
involves some unafe but race is not critical anyhow. And rollover probably
wasn't an issue either

+++ firmware/src/time.rs
@@ -2,7 +2,7 @@
 use core::fmt;
 
 #[derive(Copy, Clone, Debug)]
-pub struct SimpleTime(pub u32);
+pub struct SimpleTime(pub i64);
 
 impl fmt::Display for SimpleTime {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
